Safari-609.1.11.1/LayoutTests/fast/dom/Node/DOMNodeRemovedEvent.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script src="../../../resources/js-test-pre.js"></script>
</head>
<body>
<script>
description("This test checks that DOMNodeRemovedEvent is emitted once (and only once).");

var div = document.createElement("div");
document.body.appendChild(div);

var count = 0;
document.body.addEventListener("DOMNodeRemoved", function () { count++; }, false);
document.body.removeChild(div);

shouldBe("count", "1");
</script>
<script src="../../../resources/js-test-post.js"></script>
</body>
</html>
